Return exceptions to caller, check if tag exists
